+++ src/extension.ts
@@ -74,16 +74,19 @@ const getCommentEnding = (editor: vscode.TextEditor | undefined) => {
 	}
 }
 
-const sendQueryToChatGPT = async (
+const sendQueryToOpenAI = async (
 	queryText: string,
 	openAIKey: string,
 	commentClosing?: RegExp,
 ) => {
 	// get user or workspace configuration object
 	let config = vscode.workspace.getConfiguration()
+	const model = config.get("chatgpt-helper.model") as string | null
+	const modelId = model === "chatgpt" ? "gpt-3.5-turbo" : "code-davinci-002"
+	const modelNameReadable = model === "chatgpt" ? "ChatGPT" : "Codex"
 
 	const outputDocument = await vscode.workspace.openTextDocument({
-		content: "Querying ChatGPT. This may take some time...",
+		content: `Querying ${modelNameReadable}. This may take some time...`,
 		language: "markdown",
 	})
 	const outputDocumentEditor = await vscode.window.showTextDocument(
@@ -101,7 +104,8 @@ const sendQueryToChatGPT = async (
 	const request = https.request(
 		{
 			hostname: "api.openai.com",
-			path: "/v1/completions",
+			path:
+				model === "codex" ? "/v1/completions" : "/v1/chat/completions",
 			method: "POST",
 			headers: {
 				"Content-Type": "application/json", // eslint-disable-line @typescript-eslint/naming-convention
@@ -111,24 +115,48 @@ const sendQueryToChatGPT = async (
 		response => {
 			response.on("data", data => {
 				process.stdout.write(data)
-				responseText += JSON.parse(
-					data.toString(),
-				).choices[0].text.trim()
-				if (commentClosing) {
-					responseText = responseText.replace(commentClosing, "")
+				if (model === "codex") {
+					responseText += JSON.parse(
+						data.toString(),
+					).choices[0].text.trim()
+					if (commentClosing) {
+						responseText = responseText.replace(commentClosing, "")
+					}
+				} else {
+					responseText += JSON.parse(data.toString()).choices[0]
+						.message.content
 				}
 			})
 		},
 	)
 
 	request.write(
-		JSON.stringify({
-			model: "code-davinci-002",
-			prompt: entireQueryText,
-			temperature: 0.1,
-			max_tokens: 512, // eslint-disable-line @typescript-eslint/naming-convention
-			frequency_penalty: 0.38, // eslint-disable-line @typescript-eslint/naming-convention
-		}),
+		JSON.stringify(
+			model === "codex"
+				? {
+						model: "code-davinci-002",
+						prompt: entireQueryText,
+						temperature: 0.1,
+						max_tokens: 512, // eslint-disable-line @typescript-eslint/naming-convention
+						frequency_penalty: 0.38, // eslint-disable-line @typescript-eslint/naming-convention
+				  }
+				: {
+						model: "gpt-3.5-turbo",
+						max_tokens: 512, // eslint-disable-line @typescript-eslint/naming-convention
+						// todo: should subtract length (in tokens) of query
+						messages: [
+							{
+								role: "system",
+								content:
+									"You are a knowledgeable and helpful assistant called ChatGPT, who is specialised to help answer queries about users' code.",
+							},
+							{
+								role: "user",
+								content: entireQueryText,
+							},
+						],
+				  },
+		),
 	)
 	request.end()
 	request.on("close", () => {
@@ -151,7 +179,7 @@ const sendQueryToChatGPT = async (
 					new vscode.Position(0, 0),
 					new vscode.Position(99999999999999, 0),
 				),
-				`Error querying ChatGPT.\n\n${error}`,
+				`Error querying ${modelNameReadable}.\n\n${error}`,
 			)
 		})
 	})
@@ -205,13 +233,21 @@ export function activate(context: vscode.ExtensionContext) {
 				const commentEnding = getCommentEnding(
 					vscode.window.activeTextEditor,
 				)
-				sendQueryToChatGPT(
-					codeToQuery +
-						"\n" +
-						commentOpening +
-						(workspaceConfiguration.get(
-							"chatgpt-helper.codex.queries.codeNotWorking",
-						) as string | null),
+				const model = workspaceConfiguration.get(
+					"chatgpt-helper.model",
+				) as string
+				const fullQueryText =
+					model === "chatgpt"
+						? "Why is the following code not working?\n" +
+						  codeToQuery
+						: codeToQuery +
+						  "\n" +
+						  commentOpening +
+						  (workspaceConfiguration.get(
+								"chatgpt-helper.codex.queries.codeNotWorking",
+						  ) as string | null)
+				sendQueryToOpenAI(
+					fullQueryText,
 					await getOpenAIKey(),
 					commentEnding,
 				)
@@ -249,13 +285,20 @@ export function activate(context: vscode.ExtensionContext) {
 				const commentEnding = getCommentEnding(
 					vscode.window.activeTextEditor,
 				)
-				sendQueryToChatGPT(
-					codeToQuery +
-						"\n" +
-						commentOpening +
-						(workspaceConfiguration.get(
-							"chatgpt-helper.codex.queries.explainCode",
-						) as string | null),
+				const model = workspaceConfiguration.get(
+					"chatgpt-helper.model",
+				) as string
+				const fullQueryText =
+					model === "chatgpt"
+						? "Explain the following code:\n" + codeToQuery
+						: codeToQuery +
+						  "\n" +
+						  commentOpening +
+						  (workspaceConfiguration.get(
+								"chatgpt-helper.codex.queries.explainCode",
+						  ) as string | null)
+				sendQueryToOpenAI(
+					fullQueryText,
 					await getOpenAIKey(),
 					commentEnding,
 				)
@@ -276,7 +319,7 @@ export function activate(context: vscode.ExtensionContext) {
 				placeHolder: "Query",
 			})
 			if (queryText) {
-				sendQueryToChatGPT(queryText, await getOpenAIKey())
+				sendQueryToOpenAI(queryText, await getOpenAIKey())
 			} else {
 				vscode.window.showErrorMessage(
 					"No query entered. Did not send to ChatGPT",